+++ CHANGELOG.md
@@ -5,14 +5,48 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.17.0
+
+### Feature
+
+- Workflow : Display by default is now tasks only
+
+
+## 0.16.3
+
+### Fix
+
+- Dataset : Update plot to new decorator paradigm
+
+
+## 0.16.2
+
+### Fix
+
+- Export: Add block_index if export blocks
+- PhysicalObject : Remove to_stl encoding element
+- Schema : Take `None` default value into account
+- WorkflowRun : Compute memory usage, subtract before_memory
+
+
+### Build
+
+- Numpy : Restrict version to lower 2.0.0
+
 
 ## 0.16.1
 
 ### Added
 
-- Export : Define export method with decorator 
+- Export : Define export method with decorator
 
-### Performance
+### Fix
+
+- ModelMethod : Call the right dict_to_object of MethodType
+- MultiPlot : Add selector_name MultiPlot script
+- WorkflowRun : Add markdown to display_settings
+
+### Tests
 
 - Schema : Add new unittest to test get_import_names methods